Refactor test utilities: update import paths and enhance type definitions
Modified the import paths in various test files to ensure consistency and clarity by referencing the updated test utilities. Additionally, enhanced type definitions in the custom render function to improve type safety and maintainability across the testing suite.
