+++ packages/ui/src/components/__tests__/ExitIntentModal.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { render, screen, fireEvent, waitFor } from "../utils/test-utils";
-import { ExitIntentModal } from "../components/ExitIntentModal";
+import { fireEvent, render, screen, waitFor } from "../../__tests__/utils/test-utils";
+import { ExitIntentModal } from "../ExitIntentModal";
 
 // Mock createPortal for testing
 jest.mock("react-dom", () => ({
